app4prod/requirements.txt
asgiref==3.7.2
Django==4.2.7
gunicorn==21.2.0
packaging==23.2
sqlparse==0.4.4
typing_extensions==4.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
whitenoise==6.6.0